+++ 1.Information-Gathering/1.Network_Portscanner/thanos/thanos-update.sh
@@ -3,6 +3,28 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/P0cL4bs/Thanos.git
+GITREPOROOT=/opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/thanos/P0cL4bs/Thanos
+GITCONFDIR=/opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/thanos/P0cL4bs/Thanos/.git
+GITCLONEDIR=/opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/thanos/P0cL4bs
+EXECUTEABLE1=thanos	
+EXECUTEABLE2=thanos
+BINDIR=/usr/local/bin
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/1.Network_Portscanner
+DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/1.Network_Portscanner
+DSKTPFL=thanos.desktop
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/master
+	git pull
+}
+
 echo "${bold}
  _____ _   _    _    _   _  ___  ____  
 |_   _| | | |  / \  | \ | |/ _ \/ ___| 
@@ -10,59 +32,57 @@ echo "${bold}
   | | |  _  |/ ___ \| |\  | |_| |___) |
   |_| |_| |_/_/   \_\_| \_|\___/|____/ 
         
+UPDATE CHECK
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/thanos/P0cL4bs/Thanos
-GITREPOGITFILE=$GITREPOROOT/.git
-EXECUTEABLE1=thanos	
-EXECUTEABLE2=thanos
-
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/1.Network_Portscanner
-DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/1.Network_Portscanner
-DSKTPFL=thanos.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/thanos/P0cL4bs
-cd /opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/thanos/P0cL4bs
-git clone https://github.com/P0cL4bs/Thanos.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
-
+if git checkout master &&
+    git fetch origin master &&
+    [ `git rev-list HEAD...origin/master --count` != 0 ] &&
+    git merge origin/master
+then
+    
 cd $GITREPOROOT
 #make clean
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init 
-git submodule update --recursive
+GITRESET
+GITSBMDLINIT
 #
 make -j 4
 #sudo make install
 
-sudo rm /usr/local/bin/$EXECUTEABLE2
 chmod +x $GITREPOROOT/$EXECUTEABLE1
+sudo rm -f $BINDIR/$EXECUTEABLE2
 sudo ln -s $GITREPOROOT/$EXECUTEABLE1 /usr/local/bin/$EXECUTEABLE2
+
 mkdir -p $DSKTPFLSDEST
 rm -f $DSKTPFLSDEST/$DSKTPFL
-
-
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
 thanos --new-db banner_grabber.db
+
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
