+++ 1.Information-Gathering/1.Network_Portscanner/zmap/zmap-update.sh
@@ -3,6 +3,25 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/zmap/zmap.git
+GITREPOROOT=/opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/zmap/zmap/zmap
+GITCONFDIR=/opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/zmap/zmap/zmap/.git
+GITCLONEDIR=/opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/zmap/zmap
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/1.Network_Portscanner
+DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/1.Network_Portscanner
+DSKTPFL=zmap.desktop
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/master
+	git pull
+}
+
 echo "${bold}
  _______  __    _    ____  
 |__  /  \/  |  / \  |  _ \ 
@@ -10,59 +29,60 @@ echo "${bold}
  / /_| |  | |/ ___ \|  __/ 
 /____|_|  |_/_/   \_\_|    
             
+UPDATE CHECK
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/zmap/zmap/zmap
-GITREPOGITFILE=$GITREPOROOT/.git
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/1.Network_Portscanner
-DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/1.Network_Portscanner
-DSKTPFL=zmap.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/zmap/zmap
-cd /opt/ITSEC/1.Information-Gathering/1.Network_Portscanner/zmap/zmap
-git clone https://github.com/zmap/zmap.git
-
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
+if git checkout master &&
+    git fetch origin master &&
+    [ `git rev-list HEAD...origin/master --count` != 0 ] &&
+    git merge origin/master
+then
+    
 cd $GITREPOROOT
 
 cd build
 make clean
 cd $GITREPOROOT
 rm -f build
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
-#
+GITRESET
+GITSBMDLINIT
+
 mkdir build
 cd build
 cmake ..
 make -j 4
-sudo make install 
-rm -f  $DSKTPFLSDEST/$DSKTPFL
+sudo make install
+
 mkdir -p $DSKTPFLSDEST
+rm -f $DSKTPFLSDEST/$DSKTPFL
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
 
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
