+++ 1.Information-Gathering/13.OSINT/tekdefense-automater/tekdefense-automater-update.sh
@@ -3,59 +3,80 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/1aN0rmus/TekDefense-Automater.git
+GITREPOROOT=/opt/ITSEC/1.Information-Gathering/13.OSINT/tekdefense-automater/1aN0rmus/TekDefense-Automater
+GITCONFDIR=/opt/ITSEC/1.Information-Gathering/13.OSINT/tekdefense-automater/1aN0rmus/TekDefense-Automater/.git
+GITCLONEDIR=/opt/ITSEC/1.Information-Gathering/13.OSINT/tekdefense-automater/1aN0rmus
+EXECUTEABLE1=Automater.py
+EXECUTEABLE2=automater
+BINDIR=/usr/local/bin
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/13.OSINT
+DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/13.OSINT
+DSKTPFL=automater.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/master
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
     _   _   _ _____ ___  __  __    _  _____ _____ ____  
    / \ | | | |_   _/ _ \|  \/  |  / \|_   _| ____|  _ \ 
   / _ \| | | | | || | | | |\/| | / _ \ | | |  _| | |_) |
  / ___ \ |_| | | || |_| | |  | |/ ___ \| | | |___|  _ < 
 /_/   \_\___/  |_| \___/|_|  |_/_/   \_\_| |_____|_| \_\
-          
+UPDATE  
 ${normal}"
 
-
-GITREPOROOT=/opt/ITSEC/1.Information-Gathering/13.OSINT/tekdefense-automater/1aN0rmus/TekDefense-Automater
-GITREPOGITFILE=$GITREPOROOT/.git
-EXECUTEABLE1=Automater.py
-EXECUTEABLE2=automater
-
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/13.OSINT
-DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/13.OSINT
-DSKTPFL=automater.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-
-mkdir -p /opt/ITSEC/1.Information-Gathering/13.OSINT/tekdefense-automater/1aN0rmus
-cd /opt/ITSEC/1.Information-Gathering/13.OSINT/tekdefense-automater/1aN0rmus
-git clone https://github.com/1aN0rmus/TekDefense-Automater.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
+if git checkout master &&
+    git fetch origin master &&
+    [ `git rev-list HEAD...origin/master --count` != 0 ] &&
+    git merge origin/master
+then
+    
 cd $GITREPOROOT
-sudo rm -f /usr/local/bin/$EXECUTEABLE2
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
+
+GITRESET
+
+GITSBMDLINIT
+
 chmod +x $GITREPOROOT/$EXECUTEABLE1
+sudo rm -f /usr/local/bin/$EXECUTEABLE2
 sudo ln -s $GITREPOROOT/$EXECUTEABLE1 /usr/local/bin/$EXECUTEABLE2
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
