+++ 1.Information-Gathering/13.OSINT/theharvester/theharvester-update.sh
@@ -3,63 +3,85 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/laramies/theHarvester.git
+GITREPOROOT=/opt/ITSEC/1.Information-Gathering/13.OSINT/theharvester/laramies/theHarvester
+GITCONFDIR=/opt/ITSEC/1.Information-Gathering/13.OSINT/theharvester/laramies/theHarvester/.git
+GITCLONEDIR=/opt/ITSEC/1.Information-Gathering/13.OSINT/theharvester/laramies
+EXECUTEABLE1=theHarvester.py
+EXECUTEABLE2=theharvester
+EXECUTEABLE3=myparser.py
+EXECUTEABLE4=myparser
+BINDIR=/usr/local/bin
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/13.OSINT
+DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/13.OSINT
+DSKTPFL=theharvester.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/master
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
  _____ _   _ _____ _   _    _    ______     _______ ____ _____ _____ ____  
 |_   _| | | | ____| | | |  / \  |  _ \ \   / / ____/ ___|_   _| ____|  _ \ 
   | | | |_| |  _| | |_| | / _ \ | |_) \ \ / /|  _| \___ \ | | |  _| | |_) |
   | | |  _  | |___|  _  |/ ___ \|  _ < \ V / | |___ ___) || | | |___|  _ < 
   |_| |_| |_|_____|_| |_/_/   \_\_| \_\ \_/  |_____|____/ |_| |_____|_| \_\
-    
+INSTALL
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/1.Information-Gathering/13.OSINT/theharvester/laramies/theHarvester
-GITREPOGITFILE=$GITREPOROOT/.git
-EXECUTEABLE1=theHarvester.py
-EXECUTEABLE2=theharvester
-EXECUTEABLE3=myparser.py
-EXECUTEABLE4=myparser
-
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/13.OSINT
-DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/13.OSINT
-DSKTPFL=theharvester.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/1.Information-Gathering/13.OSINT/theharvester/laramies
-cd /opt/ITSEC/1.Information-Gathering/13.OSINT/theharvester/laramies
-git clone https://github.com/laramies/theHarvester.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
+if git checkout master &&
+    git fetch origin master &&
+    [ `git rev-list HEAD...origin/master --count` != 0 ] &&
+    git merge origin/master
+then
+    
 cd $GITREPOROOT
-sudo rm -f /usr/local/bin/$EXECUTEABLE2
-sudo rm -f /usr/local/bin/$EXECUTEABLE4
-sudo updatedb
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
+
+GITRESET
+
+GITSBMDLINIT
+
 chmod +x $GITREPOROOT/$EXECUTEABLE1
-sudo ln -s $GITREPOROOT/$EXECUTEABLE1 /usr/local/bin/$EXECUTEABLE2
+sudo rm -f $BINDIR/$EXECUTEABLE2
+sudo rm -f $BINDIR/$EXECUTEABLE4
+sudo ln -s $GITREPOROOT/$EXECUTEABLE1 $BINDIR/$EXECUTEABLE2
 chmod +x $GITREPOROOT/$EXECUTEABLE3
-sudo ln -s $GITREPOROOT/$EXECUTEABLE3 /usr/local/bin/$EXECUTEABLE4
+sudo ln -s $GITREPOROOT/$EXECUTEABLE3 $BINDIR/$EXECUTEABLE4
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
