+++ 1.Information-Gathering/15.Fuzzer/dotdotpwn/dotdotpwn-update.sh
@@ -3,61 +3,84 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/wireghoul/dotdotpwn.git
+GITREPOROOT=/opt/ITSEC/1.Information-Gathering/15.Fuzzer/dotdotpwn/wireghoul/dotdotpwn
+GITCONFDIR=/opt/ITSEC/1.Information-Gathering/15.Fuzzer/dotdotpwn/wireghoul/dotdotpwn/.git
+GITCLONEDIR=/opt/ITSEC/1.Information-Gathering/15.Fuzzer/dotdotpwn/wireghoul
+EXECUTEABLE1=dotdotpwn.pl
+EXECUTEABLE2=dotdotpwn
+BINDIR=/usr/local/bin
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/15.Fuzzer
+DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/15.Fuzzer
+DSKTPFL=dotdotpwn.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/master
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
+
 echo "${bold}
  ____   ___ _____ ____   ___ _____ ______        ___   _ 
 |  _ \ / _ \_   _|  _ \ / _ \_   _|  _ \ \      / / \ | |
 | | | | | | || | | | | | | | || | | |_) \ \ /\ / /|  \| |
 | |_| | |_| || | | |_| | |_| || | |  __/ \ V  V / | |\  |
 |____/ \___/ |_| |____/ \___/ |_| |_|     \_/\_/  |_| \_|
-           
+UPDATE
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/1.Information-Gathering/15.Fuzzer/dotdotpwn/wireghoul/dotdotpwn
-GITREPOGITFILE=$GITREPOROOT/.git
-EXECUTEABLE1=dotdotpwn.pl
-EXECUTEABLE2=dotdotpwn
-
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/15.Fuzzer
-DSKTPFLSDEST=/home/$USER/.local/share/applications/1.Information-Gathering/15.Fuzzer
-DSKTPFL=dotdotpwn.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/1.Information-Gathering/15.Fuzzer/dotdotpwn/wireghoul
-cd /opt/ITSEC/1.Information-Gathering/15.Fuzzer/dotdotpwn/wireghoul
-git clone https://github.com/wireghoul/dotdotpwn.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
-sudo rm -f /usr/local/bin/$EXECUTEABLE2
-
+if git checkout master &&
+    git fetch origin master &&
+    [ `git rev-list HEAD...origin/master --count` != 0 ] &&
+    git merge origin/master
+then
+    
 cd $GITREPOROOT
-git clean -f 
-git pull
-git submodule init
-git submodule update --recursive
+
+GITRESET
+
+GITSBMDLINIT
 
 echo '#!/bin/bash
 cd /opt/ITSEC/1.Information-Gathering/15.Fuzzer/dotdotpwn/wireghoul/dotdotpwn
 perl dotdotpwn.pl "$@"' > $EXECUTEABLE2
 chmod +x $GITREPOROOT/$EXECUTEABLE1
 chmod +x $GITREPOROOT/$EXECUTEABLE2
-sudo ln -s $GITREPOROOT/$EXECUTEABLE2 /usr/local/bin/$EXECUTEABLE2
+sudo ln -s $GITREPOROOT/$EXECUTEABLE2 $BINDIR/$EXECUTEABLE2
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
